Adding the foundations for unit tests for the script checkin-test.py
Bugs: 4654

This is an intermediate commit that adds the fundamental capability to
intercept calls to the runSysCmnd(...) (and therefore echoRunSysCmnd) and
getCmndOutput(...) functions to return mock values for the return code and
error output.

This will allow me to write very strong but fast unit tests for the
checkin-test.py and similar scripts.