Fixed bug in Epetra LinearProblem unit test for copy constructor
Previously, this test looked at the original LinearProblem object and
the copied LinearProblem object and checked whether the id() of the
RHS vectors and the LHS vectors was the same.  The proper check is to
see if the data attribute (a buffer object) of the array attribute (a
numpy array) of the two vectors is the same.