Add extended-precision (QD) support to Tifpack.
Tifpack now has an optional dependence on the QD library for extended-precision
floating-point types. Ran builds/tests to ensure that Tifpack can still build and
run with and without the QD library enabled.
Added unit-test cases for the Relaxation and ILUT preconditioners with the
dd_real type.
Added an extended-precision test-driver to solve a linear-system with belos.
This is not yet working. (i.e., the solve isn't working.) More investigation required.