Added experimental KL expansion capbilities.
1) Added an analytic solution to the KL eigenproblem for the exponential
covariance kernel on a rectangular domain.  This will be useful for
putting KL expansions in codes.  Some unit tests were also added, plus
an example.

2) Added tools to build KL expansions from PC expansions using Anasazi.
This added an optional dependency of Stokhos on Anasazi.

3) Added an experimental matrix-free epetra operator based on taking the
KL expansion of the PC Jacobian.

4) Added a builder for building a Gram-Schmidt-Stieltjes basis from a PC
expansion, simplifying using this technique.

5) Changed the SGModelEvaluator to take an OrthogPolyApprox for
specifying the initial parameter expansion instead of an array of an
array of rcp's to Epetra_Vectors.

6) Changed the Legendre basis to use Gauss-Legendre points for sparse grids,
and added a new basis called ClenshawCurtisLegendreBasis for using
Clenshaw-Curtis points.

7) Added Ifpack and ML preconditioner factories.  This adds an optional
dependency on Ifpack.  I removed the corresponding preconditioner factory code
out of the PCE example.  Also fixed up the labels of the Epetra operators, and
fixed some strings without namespace qualification.

8) Added a 1-D exponential random field to FEApp.

9) Added some examples that experiment with new KL capabilities.