Changed over to use newer boost 1.40.0 and not just whatever is in /usr/include
In order to get consistent timing results for boost::shared_ptr across
machines, I installed boost 1.40.0 which, as of the time of this
commit, was only one version old (the current is 1.41.1).

The performance of boost::shared_ptr seems to have gone down hill with
the 1.40.0 release as opposed to whatever was in /usr/include/boost on
godel before.  This should make the performance tests pass that failed
last night.