Finally fixing the SERIAL_PERF configure
When you specify multiple arguments to CMAKE_CXX_FLAGS, you need to
use '\ ' for spaces.  I tested this on godel itself and it seems to
work perfectly.  Now boost thread support should be disabled to give
me the timing results that I want.

This should fix the configure failure:

    http://trilinos.sandia.gov/cdash/viewConfigure.php?buildid=177026

and the one the morning before.