Fixed the buildReorderedLinearOp function. It now works for purely block diagonal matrices.
Previously it crashed whenever for these matrices.  Essentially off diagonal
components were Null and this caused problems because the vector space of some
of the composite blocks was not defined.  To fixed this, the function now replaces
any Null matrix with a ZeroLinearOp.  This seems to fix the problem.

Also added a function to reorder a vector space.  Currently its not
needed or used (or tested).