Update to the cmake equivalent to the makefile export system
Added a few new variables to the existing TrilinosConfig.cmake file. These
include the compile options and whether or not the build was a shared build or
not.

Also modified the system to create a version of TrilinosConfig.cmake that will
be installed when make install is run. This version will have all the paths be
based on the install directory instead of the build directory so there will be
no dependence on the build directory. This version also gets the benefit of
having all of the libraries and their dependencies hooked up automatically. This
is through the use of the cmake command install(export...) This creates a new
set of files that declare and add all the configured libraries as well as set
their properties for dependencies/location/etc.

There is a gotcha here though. There appears to be a potential bug in cmake that
causes these generated files to over strip the path that the libraries are in
causing the path to be invalid. There is a simple workaround for now, though it
requires manually editting the generated files. The file
"Trilinos-targets-<build type>.cmake" needs to have the second
"GET_FILENAME_COMPONENT" commented out. Once that is done the correct location
of libraries is set and they can be easily added as dependencies for other
packages. Kitware has been made aware of this issue.

Unfortunately I could not find a corrolary for the install(export...) command
for the in build tree TrilinosConfig.cmake, but since this is not the typical
way we want users to depend on Trilinos I don't see that as a huge problem.
Steps were taken to make sure that the TrilinosConfig.cmake file that is
generated for the build tree still provides all the same features and doesn't
have any errors because of the "missing" Trilinos-targets files.