Disabling all Thyra dense solve tests on Darwin for now
BUGS:  4699

Here I have disabled all of the dense solve-based tests.

I think the root of the problem is a bad algorithm for seeding the random
number generator.  I put something in to set the seeds based on column indexes
to make them the same across processes and I think that results in columns in
the generated multi-vector to be linearly-dependent.  That results in huge
errors in the linear solve.  I will have to test this hypothesis by printing
the matrix and running octave/matlab and computing the condition number (which
I suspect is 10+16 or worse).

I will log on to a Mac/Darwin machine and try to get this fixed once and for
all.

Testing: Since I have only changed a test-related file, I am only testing
Thyra.

=============================
Build/Test Cases Summary
=============================

Tue Feb  9 11:23:08 MST 2010

Enabled Packages: Thyra

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=37,notpassed=0
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=39,notpassed=0

Local commits for this build/test group:
----------------------------------------
7c70ed9 Disabling all Thyra dense solve tests on Darwin for now