Adding the ability to call find_package for Trilinos and its packages
This system supercedes the previous system as there were several issues that
weren't able to be worked around and it ended up being not generally useful.

This new system is a CMake variable only system. What that means is that we
create a few files that will be read in by cmake when find_package is called
that set several variables that will be of use to the client project.

This system provides a find_package ability for both Trilinos as a whole and for
individual packages. The variables are all namespaced with either Trilinos or
the package name as it appears in TrilinosPackages.cmake. It is also set up to
generate these *Config.cmake files for both an installation tree and a build
tree. This way client projects can use a stable release installation or a
development build that has new work happening on a regular basis, or anything
inbetween.

The full documentation on how to use this and what the variables are can be
found in cmake/README.Find_Trilinos

A change made to the FindTPL dependecy logic as there was a bug causing it to
put the same header path on the list multiple times if there were multiple
headers found in the same directory. The fix was to simply remove duplicates
from that list after it was completely built.