Fixed Include paths for Dakota.
TriKota and Stokhos now look into the installed
Dakota for headers, not the source tree. This
was motivated by a new dependency on a Pecos
header, which is buried in the source tree but
now installed into the same flat directory as all
other headers. Thanks ti Bill Bohnhoff for the
dakoat fix.