fix for a few files that weren't being installed properly
Teuchos_Version.hpp was not being installed. I added it to the list of header
files for Teuchos since it should be there anyway and that will make sure it is
installed with the other Teuchos headers

Didasko wasn't installing its configure headers. Since  Didasko is an example
only library these files had to be installed manually since there wasn't a
library they should be associated with.