Modifed get{Lower/Upper}TriBlocks so that you can change blocks afterwards
This is simply a default option change to those functions that
specify if "endBlockFill" is called before the linear operator is returned.
If it has not been called, it is up to the user to call it.  However
they are allowed to modify some of the blocks.

Also added some debugging information to the Teko_LU2x2PreconditionerFactory