Adjust package enable logic to avoid flood of options in CMake GUI
Here, I have made some significant changes to the Trilinos CMake system to
avoid flooding the CMake GUIs with Trilinos enable options.  Now, package
enable options PACKAGE_ENABLE_[TESTS,EXAMPLES,OTHERPACKAGE] will not appear in
the list of CMake cache varaibles unless that package is not actually enabled.
I have tested this both with the QT GUI and ccmake and it seems to be doing
the job.

Now if the CMake QT GUI could just allow you to use "grouped" options and show
or hide "advanced" options separately, we would have the problem of flooding
the user with options solved as far as I am concerned.

The big disadvantage of what I have done is to avoid setting the cache
varables for PACKAGE_ENABLE_[TESTS,EXAMPLES,OTHERPACKAGE] until *after* all of
the package enable/disable logic is finished.  The enable/disable adjustment
logic therefore works with undefined varaibles for packages that have not been
enabled.

This means that the code will be more fragile to being broken becuse if you
misspell the name of a variable it will ignore it.  Given the current CMake
language and the confusing scoping rules of local and global (cache) varibles,
there is no other way that I can see to do this.

I added a unit test to make sure that Trilinos_ENABLE_Tests=ON is being
handled correctly.  All of the other enable/disable unit tests also pass.
There may be some use cases that are not covered by the unit tests but I
someone complains about things that have broken, I will need to add unit tests
for this.

Side Note: I made the majority of these changes while at an airport trying to
get Trilinos to build on MS Windows under the CMake GUI.  I was frustruated
with the flood of options so I decided to do something about it.

=============================
Build/Test Cases Summary
=============================

Mon Apr 19 14:53:36 MDT 2010

Enabled Packages: TrilinosFramework
Enabled all Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=1198,notpassed=0 (65.5239462852 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1250,notpassed=0 (40.7175182144 min)

Local commits for this build/test group:
----------------------------------------
512fa6a Adjust package enable logic to avoid flood of options in CMake GUI