Changed preconditioner from ml to ifpack in Jv example
The continuous integration server shows the newly added Jv test failing.  I could not reproduce this on my system but it seems ml was throwing an internal error due to the system being to small to apply a v cycle to.  This commit change the preconditioner to ifpack to avoid this problem.